+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The **Unreleased** section collects notes for unreleased changes and features, u
 
 ## Unreleased
 
+- [Added] server information in admin panel, and moved server restart to it
+- [Added] after login or navigate to '/' redirect to "logo url" as home page
 
 ## [7.3.191] - 2021-01-18